requirements.txt
absl-py==0.10.0
appnope==0.1.0
av==8.0.2
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@59336e152057dff1e44b605aed66c8f90af8eec5
dlib==19.21.0
future==0.18.2
fvcore==0.1.1.post20200716
google-auth==1.21.1
google-auth-oauthlib==0.4.1
grpcio==1.31.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
mock==4.0.2
networkx==2.5
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.42
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
portalocker==2.0.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
tifffile==2020.8.25
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==5.0.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
yacs==0.1.8
zipp==3.1.0
